# GREENALPHA/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GreenAlpha Carbon Footprint Engine",
    "description": "Advanced carbon footprint calculation and sustainability analytics API with <500ms response time",
    "version": "1.0.0"
  },
  "paths": {
    "/carbon/calculate": {
      "post": {
        "tags": [
          "Carbon Calculation",
          "Carbon Calculator"
        ],
        "summary": "Calculate Carbon Footprint",
        "description": "Calculate carbon footprint for a product shipment\n\nThis endpoint calculates the complete carbon footprint including:\n- Production emissions (Scope 1 & 2)\n- Transportation emissions (Scope 3)\n- Carbon cost estimation\n- Trading opportunities\n- Detailed breakdown and uncertainty analysis\n\nResponse time target: <500ms for 95% of requests",
        "operationId": "calculate_carbon_footprint_carbon_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarbonCalculationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarbonCalculationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/carbon/calculate/batch": {
      "post": {
        "tags": [
          "Carbon Calculation",
          "Carbon Calculator"
        ],
        "summary": "Calculate Batch Carbon Footprint",
        "description": "Calculate carbon footprint for multiple products in batch\n\nProcesses up to 100 calculations in parallel for optimal performance.\nReturns individual results plus optional batch summary.",
        "operationId": "calculate_batch_carbon_footprint_carbon_calculate_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCalculationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchCalculationRequest": {
        "properties": {
          "calculations": {
            "items": {
              "$ref": "#/components/schemas/CarbonCalculationRequest"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Calculations",
            "description": "List of calculations (max 100)"
          },
          "include_summary": {
            "type": "boolean",
            "title": "Include Summary",
            "description": "Include batch summary",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "calculations"
        ],
        "title": "BatchCalculationRequest",
        "description": "Batch carbon footprint calculation request"
      },
      "CarbonCalculationRequest": {
        "properties": {
          "product_name": {
            "type": "string",
            "title": "Product Name",
            "description": "Product name or category"
          },
          "quantity": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "description": "Quantity of products",
            "default": 1.0
          },
          "origin_country": {
            "type": "string",
            "title": "Origin Country",
            "description": "Origin country code (3-letter)"
          },
          "destination_country": {
            "type": "string",
            "title": "Destination Country",
            "description": "Destination country code (3-letter)"
          },
          "transport_mode": {
            "type": "string",
            "title": "Transport Mode",
            "description": "Transportation mode",
            "default": "road_truck"
          },
          "product_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Category",
            "description": "Product category",
            "default": "general"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg",
            "description": "Total weight in kg"
          },
          "origin_latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Latitude"
          },
          "origin_longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Longitude"
          },
          "destination_latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Latitude"
          },
          "destination_longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Longitude"
          },
          "custom_emission_factors": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Emission Factors",
            "description": "Custom emission factors override"
          }
        },
        "type": "object",
        "required": [
          "product_name",
          "origin_country",
          "destination_country"
        ],
        "title": "CarbonCalculationRequest",
        "description": "Carbon footprint calculation request model"
      },
      "CarbonCalculationResponse": {
        "properties": {
          "total_emissions_kg_co2e": {
            "type": "number",
            "title": "Total Emissions Kg Co2E"
          },
          "production_emissions": {
            "type": "number",
            "title": "Production Emissions"
          },
          "transportation_emissions": {
            "type": "number",
            "title": "Transportation Emissions"
          },
          "scope_1_emissions": {
            "type": "number",
            "title": "Scope 1 Emissions"
          },
          "scope_2_emissions": {
            "type": "number",
            "title": "Scope 2 Emissions"
          },
          "scope_3_emissions": {
            "type": "number",
            "title": "Scope 3 Emissions"
          },
          "carbon_cost_usd": {
            "type": "number",
            "title": "Carbon Cost Usd"
          },
          "carbon_trading_opportunities": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Carbon Trading Opportunities"
          },
          "calculation_confidence": {
            "type": "number",
            "title": "Calculation Confidence"
          },
          "response_time_ms": {
            "type": "number",
            "title": "Response Time Ms"
          },
          "calculation_method": {
            "type": "string",
            "title": "Calculation Method"
          },
          "data_sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Data Sources"
          },
          "production_breakdown": {
            "additionalProperties": true,
            "type": "object",
            "title": "Production Breakdown"
          },
          "transport_breakdown": {
            "additionalProperties": true,
            "type": "object",
            "title": "Transport Breakdown"
          },
          "uncertainty_analysis": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Uncertainty Analysis"
          },
          "recommendations": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendations"
          }
        },
        "type": "object",
        "required": [
          "total_emissions_kg_co2e",
          "production_emissions",
          "transportation_emissions",
          "scope_1_emissions",
          "scope_2_emissions",
          "scope_3_emissions",
          "carbon_cost_usd",
          "carbon_trading_opportunities",
          "calculation_confidence",
          "response_time_ms",
          "calculation_method",
          "data_sources",
          "production_breakdown",
          "transport_breakdown",
          "uncertainty_analysis"
        ],
        "title": "CarbonCalculationResponse",
        "description": "Carbon footprint calculation response model"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}